split token editor into separate file
