+++ apps/designer/app/designer/shared/design-tokens-manager/design-tokens-manager.tsx
@@ -1,21 +1,6 @@
-import { useState, useEffect, type FormEvent, useRef } from "react";
-import {
-  Button,
-  Flex,
-  InputErrorsTooltip,
-  Label,
-  List,
-  ListItem,
-  Popover,
-  PopoverContent,
-  PopoverHeader,
-  PopoverPortal,
-  PopoverTrigger,
-  TextArea,
-  TextField,
-  useList,
-} from "@webstudio-is/design-system";
-import { CheckIcon, PlusIcon } from "@webstudio-is/icons";
+import { useState } from "react";
+import { Flex, List, ListItem, useList } from "@webstudio-is/design-system";
+import { CheckIcon } from "@webstudio-is/icons";
 import type { DesignToken } from "@webstudio-is/design-tokens";
 import { designTokensGroups } from "@webstudio-is/design-tokens";
 import { useDesignTokens } from "~/shared/nano-states";
@@ -23,14 +8,10 @@ import type { Publish } from "~/shared/pubsub";
 // @todo this is temporary, we need to either make that collapsible reusable or copy it over
 // This wasn't properly designed, so this is mostly temp
 import { CollapsibleSection } from "../inspector";
-import {
-  deleteTokenMutable,
-  filterByType,
-  findByName,
-  updateTokenMutable,
-} from "./utils";
+import { deleteTokenMutable, filterByType, updateTokenMutable } from "./utils";
 import { useMenu } from "./item-menu";
 import produce from "immer";
+import { type DesignTokenSeed, TokenEditor } from "./token-editor";
 
 declare module "~/shared/pubsub" {
   export interface PubsubMap {
@@ -43,164 +24,6 @@ declare module "~/shared/pubsub" {
     deleteToken: DesignToken["name"];
   }
 }
-
-const validate = (
-  tokens: Array<DesignToken>,
-  data: Partial<DesignToken>,
-  isNew: boolean
-): { name: Array<string>; value: Array<string>; hasErrors: boolean } => {
-  const name = [];
-  const value = [];
-
-  if (String(data.name).trim() === "") name.push("Name is required");
-  if (isNew && findByName(tokens, data?.name)) {
-    name.push("Name is already taken");
-  }
-  if (String(data.value).trim() === "") value.push("Value is required");
-
-  return {
-    name,
-    value,
-    hasErrors: name.length !== 0 || value.length !== 0,
-  };
-};
-
-const initialErrors = {
-  name: [],
-  value: [],
-  hasErrors: false,
-};
-
-const getToken = (
-  form: HTMLFormElement,
-  tokenOrSeed?: DesignToken | DesignTokenSeed
-) => {
-  const formData = new FormData(form);
-  const data = Object.fromEntries(formData);
-  return { ...tokenOrSeed, ...data } as DesignToken;
-};
-
-type DesignTokenSeed = Pick<DesignToken, "group" | "type">;
-
-const TokenEditor = ({
-  token,
-  seed,
-  trigger,
-  isOpen,
-  onChangeComplete,
-  onOpenChange,
-}: {
-  token?: DesignToken;
-  seed?: DesignTokenSeed;
-  trigger?: JSX.Element;
-  isOpen: boolean;
-  onChangeComplete: (token: DesignToken) => void;
-  onOpenChange: (isOpen: boolean) => void;
-}) => {
-  const [tokens] = useDesignTokens();
-  const [errors, setErrors] =
-    useState<ReturnType<typeof validate>>(initialErrors);
-  const formRef = useRef<HTMLFormElement>(null);
-
-  useEffect(() => {
-    if (isOpen === false && errors.hasErrors) {
-      setErrors(initialErrors);
-    }
-  }, [isOpen, errors.hasErrors]);
-
-  const handleChange = (event: FormEvent<HTMLFormElement>) => {
-    if (errors.hasErrors === false || formRef.current === null) {
-      return;
-    }
-    const updatedToken = getToken(formRef.current, token ?? seed);
-    const nextErrors = validate(tokens, updatedToken, token === undefined);
-    setErrors(nextErrors);
-  };
-
-  const handleSubmit = (event?: FormEvent<HTMLFormElement>) => {
-    event?.preventDefault();
-    if (formRef.current === null) return;
-    const updatedToken = getToken(formRef.current, token ?? seed);
-    const nextErrors = validate(tokens, updatedToken, token === undefined);
-    setErrors(nextErrors);
-
-    if (nextErrors.hasErrors === false) {
-      onChangeComplete(updatedToken);
-      onOpenChange(false);
-    }
-  };
-
-  const handleOpenChange = (isOpen: boolean) => {
-    if (isOpen === false) {
-      handleSubmit();
-      return;
-    }
-    onOpenChange(isOpen);
-  };
-
-  return (
-    <Popover modal open={isOpen} onOpenChange={handleOpenChange}>
-      <PopoverTrigger
-        asChild
-        aria-label={token === undefined ? "Create Token" : "Edit Token"}
-      >
-        {trigger ?? (
-          <Button
-            ghost
-            onClick={(event) => {
-              event.preventDefault();
-              onOpenChange(true);
-            }}
-          >
-            <PlusIcon />
-          </Button>
-        )}
-      </PopoverTrigger>
-      <PopoverPortal>
-        <PopoverContent align="end" css={{ zIndex: "$zIndices$1" }}>
-          <form onChange={handleChange} onSubmit={handleSubmit} ref={formRef}>
-            <Flex direction="column" gap="2" css={{ padding: "$spacing$7" }}>
-              <Label htmlFor="name">Name</Label>
-              <InputErrorsTooltip
-                errors={errors.name}
-                css={{ zIndex: "$zIndices$2" }}
-              >
-                <TextField
-                  id="name"
-                  name="name"
-                  defaultValue={token?.name ?? ""}
-                />
-              </InputErrorsTooltip>
-              <Label htmlFor="value">Value</Label>
-              <InputErrorsTooltip
-                errors={errors.value}
-                css={{ zIndex: "$zIndices$2" }}
-              >
-                <TextField
-                  id="value"
-                  name="value"
-                  defaultValue={token?.value ?? ""}
-                />
-              </InputErrorsTooltip>
-              <Label htmlFor="description">Description</Label>
-              <TextArea
-                id="description"
-                name="description"
-                defaultValue={token?.description ?? ""}
-              />
-              {token === undefined && (
-                <Button type="submit" variant="blue">
-                  Create
-                </Button>
-              )}
-            </Flex>
-          </form>
-          <PopoverHeader title={token?.name ?? "New Token"} />
-        </PopoverContent>
-      </PopoverPortal>
-    </Popover>
-  );
-};
 
 export const DesignTokensManager = ({ publish }: { publish: Publish }) => {
   const [tokens, setTokens] = useDesignTokens();
@@ -230,7 +53,6 @@ export const DesignTokensManager = ({ publish }: { publish: Publish }) => {
       setEditingToken(tokens[index]);
     },
   });
-  const listProps = getListProps();
 
   const renderTokenEditor = ({
     token,
@@ -276,6 +98,7 @@ export const DesignTokensManager = ({ publish }: { publish: Publish }) => {
   };
 
   let index = -1;
+  const listProps = getListProps();
 
   return (
     <List
@@ -301,7 +124,7 @@ export const DesignTokensManager = ({ publish }: { publish: Publish }) => {
             <Flex direction="column">
               {filterByType(tokens, type).map((token) => {
                 const itemProps = getItemProps({ index: ++index });
-                const listItem = (
+                return (
                   <ListItem
                     {...itemProps}
                     key={token.name}
@@ -323,7 +146,6 @@ export const DesignTokensManager = ({ publish }: { publish: Publish }) => {
                     })}
                   </ListItem>
                 );
-                return listItem;
               })}
             </Flex>
           </CollapsibleSection>
